Renamed "g2d.rs" to "back_end.rs"
Closes https://github.com/PistonDevelopers/gfx_graphics/issues/148
