+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 
 name = "piston2d-gfx_graphics"
-version = "0.1.5"
+version = "0.1.6"
 authors = ["bvssvni <bvssvni@gmail.com>"]
 keywords = ["graphics", "2d", "gfx"]
 description = "A Gfx 2D back-end for the Piston game engine"
