+++ src/lib.rs
@@ -12,11 +12,11 @@ extern crate freetype;
 
 pub use gfx_texture::Texture;
 
-pub use g2d::Gfx2d;
-pub use g2d::GfxGraphics;
-pub use g2d::Gfx2d as G2D;
-pub use g2d::GfxGraphics as GraphicsBackEnd;
+pub use back_end::Gfx2d;
+pub use back_end::GfxGraphics;
+pub use back_end::Gfx2d as G2D;
+pub use back_end::GfxGraphics as GraphicsBackEnd;
 pub use glyph::GlyphCache;
 
-mod g2d;
+mod back_end;
 mod glyph;
